[#235] :recycle: refactor: useAffixedSticker hook 분리
