+++ client/src/pages/DiaryDetail.tsx
@@ -3,14 +3,12 @@ import { useParams, useLocation } from "react-router-dom";
 import { useSetRecoilState, useResetRecoilState } from "recoil";
 import Fade from "react-reveal/Fade";
 
-import { AffixedStickerInfo } from "src/@types/response";
 import { focusedDiaryIdAtom, isDeleteModalVisibleAtom } from "src/recoil/diary-detail/atom";
-import { api } from "src/utils/axiosApi/api";
-import { API_URL } from "src/constants/API_URL";
-import BackButton from "src/components/common/BackButton";
+import { BackButton } from "src/components/common";
 import { StickerSaveBtn, StickerSection, DiaryContainer } from "src/components/diary-detail";
-import * as G from "src/styles/Common";
 import useNewSticker from "src/hooks/useNewSticker";
+import useAffixedSticker from "src/hooks/useAffixedSticker";
+import * as G from "src/styles/Common";
 
 const DiaryDetail = () => {
   const params = useParams();
@@ -26,28 +24,12 @@ const DiaryDetail = () => {
     handleDeleteStickers,
     handleResetSelectedStcks,
   } = useNewSticker();
+  const { stickers, getAffixedStickers, handleUpdateAffixedStickers } = useAffixedSticker(
+    Number(params.diaryId)
+  );
   const [isStickerEditing, setIsStickerEditing] = useState<boolean>(false);
   const changeStickerEditState = () => {
     setIsStickerEditing((prev) => !prev);
-  };
-  const [stickers, setStickers] = useState<AffixedStickerInfo[]>([]);
-  const getAffixedStickers = async () => {
-    try {
-      const { data } = await api.get(API_URL.stickers(Number(params.diaryId)));
-      setStickers(data.data);
-    } catch (err) {
-      if (err instanceof Error) alert(err.message);
-    }
-  };
-  const handleUpdateAffixedStickers = (newSticker: AffixedStickerInfo) => {
-    setStickers((curStickers) => {
-      return [
-        ...curStickers.filter(
-          (sticker) => sticker.stickedStickerId !== newSticker.stickedStickerId
-        ),
-        newSticker,
-      ];
-    });
   };
 
   useEffect(() => {
